Made checks for roles on super-admin page, added new features for super-admin making it easier to navigate through the pages. Moved calendar to dashboard where it should be since its only visible when logged in and has a user_id, also added a new element for the user to put his availibity in so it makes it easier to check your avaibility. Fixed an error log with a javascript file not being found anywhere apart from the index page, fixed an error on the calendar class, that caused it to be buggy. More changes too cant remember them.

+++ Private/deleteUser.php
@@ -1,4 +1,19 @@
 <?php include("../Layout/header.php");
 
-$user = (new User)->deleteUser($_GET['user_id']);
-header("location:users.php");
+if($_SESSION['role'] == 1 ){
+    echo("Uw heeft niet de rechten om dit te zien, over 5 seconden word u teruggestuurd naar uw dashboard.");
+    header("refresh:5;url=dashboard.php");
+}
+if($_SESSION['role'] == 2 ){
+    echo("Uw heeft niet de rechten om dit te zien, over 5 seconden word u teruggestuurd naar uw dashboard.");
+    header("refresh:5;url=dashboard.php");
+}
+if($_SESSION['role'] == 3) {
+    $user = (new User)->deleteUser($_GET['user_id']);
+    header("location:users.php");
+}
+
+
+
+
+
